+++ android/src/main/java/com/azifacemobile/theme/Button.java
@@ -1,16 +1,32 @@
 package com.azifacemobile.theme;
 
+import android.graphics.Typeface;
+
+import com.azifacemobile.theme.abstracts.ViewStyle;
 import com.azifacemobile.utils.Theme;
+import com.facebook.react.bridge.ReactApplicationContext;
 
 import org.json.JSONObject;
 
-public class Button {
+import javax.annotation.Nullable;
+
+public class Button extends ViewStyle {
+  private static final String KEY = "button";
   private final JSONObject theme;
+  private final Font font;
   private final Color color;
 
-  public Button(JSONObject theme) {
-    this.theme = new Theme().getTarget(theme, "button");
+  public Button(ReactApplicationContext context, JSONObject target) {
+    super(target, KEY);
+
+    this.theme = new Theme().getTarget(target, KEY);
+    this.font = new Font(context);
     this.color = new Color();
+  }
+
+  @Override
+  public int getCornerRadius() {
+    return super.getCornerRadius(-1);
   }
 
   public int getBackgroundNormalColor() {
@@ -35,5 +51,10 @@ public class Button {
 
   public int getTextHighlightColor() {
     return this.color.getColor(this.theme, "textHighlightColor");
+  }
+
+  @Nullable
+  public Typeface getFont() {
+    return this.font.getTypography(this.theme, "font");
   }
 }