+++ android/src/main/java/com/azifacemobile/theme/CaptureScreen.java
@@ -1,23 +1,35 @@
 package com.azifacemobile.theme;
 
-import com.azifacemobile.theme.abstracts.CommonStyle;
+import android.graphics.Typeface;
+
+import com.azifacemobile.theme.abstracts.ViewStyle;
 import com.azifacemobile.utils.Theme;
+import com.facebook.react.bridge.ReactApplicationContext;
 
 import org.json.JSONObject;
 
-public class CaptureScreen extends CommonStyle {
+import javax.annotation.Nullable;
+
+public class CaptureScreen extends ViewStyle {
   private static final String KEY = "captureScreen";
-  private final JSONObject theme;
+  private final JSONObject target;
+  private final Font font;
   private final Color color;
 
-  public CaptureScreen(JSONObject theme) {
-    super(theme, KEY);
+  public CaptureScreen(ReactApplicationContext context, JSONObject target) {
+    super(target, KEY);
 
-    this.theme = new Theme().getTarget(theme, KEY);
+    this.target = new Theme().getTarget(target, KEY);
+    this.font = new Font(context);
     this.color = new Color();
   }
 
   public int getFrameStrokeColor() {
-    return this.color.getColor(this.theme, "frameStrokeColor");
+    return this.color.getColor(this.target, "frameStrokeColor");
+  }
+
+  @Nullable
+  public Typeface getFont() {
+    return this.font.getTypography(this.target, "font");
   }
 }