+++ android/src/main/java/com/azifacemobile/theme/Feedback.java
@@ -1,23 +1,26 @@
 package com.azifacemobile.theme;
 
-import com.azifacemobile.theme.abstracts.CommonStyle;
+import android.graphics.Typeface;
+
+import com.azifacemobile.theme.abstracts.ViewStyle;
 import com.azifacemobile.utils.Theme;
+import com.facebook.react.bridge.ReactApplicationContext;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class Feedback extends CommonStyle {
+public class Feedback extends ViewStyle {
   private static final String KEY = "feedback";
   private final Theme theme;
   private final JSONObject target;
-  private final General general;
+  private final Font font;
   private final Color color;
 
-  public Feedback() {
+  public Feedback(ReactApplicationContext context) {
     super(KEY);
 
     this.theme = new Theme();
-    this.general = new General();
+    this.font = new Font(context);
     this.color = new Color();
 
     this.target = this.theme.getTarget(KEY);
@@ -25,19 +28,21 @@ public class Feedback extends CommonStyle {
 
   @Override
   public int getBackgroundColor() {
-    return this.getBackgroundColor("#026ff4");
+    return super.getBackgroundColor("#026ff4");
+  }
+
+  @Override
+  public int getCornerRadius() {
+    return super.getCornerRadius(-1);
+  }
+
+  @Override
+  public int getElevation() {
+    return this.getElevation(10);
   }
 
   public int getTextColor() {
     return this.color.getColor(this.target, "textColor");
-  }
-
-  public int getBorderRadius() {
-    return this.general.getBorderRadius(this.target, "cornerRadius", -1);
-  }
-
-  public int getElevation() {
-    return this.general.getElevation(this.target, "elevation", 10);
   }
 
   public boolean getEnablePulsatingText() {
@@ -51,5 +56,9 @@ public class Feedback extends CommonStyle {
     } catch (JSONException e) {
       return true;
     }
+  }
+
+  public Typeface getFont() {
+    return this.font.getTypography(this.target, "font");
   }
 }