+++ examples/expoapp/app/index.tsx
@@ -13,9 +13,13 @@ import {
   liveness,
   photoScan,
   photoMatch,
+  setLocale,
+  setDynamicStrings,
+  resetDynamicStrings,
   vocal,
   type Processor,
   type Headers,
+  type Locale,
 } from '@azify/aziface-mobile';
 import {
   getDeviceId,
@@ -27,6 +31,7 @@ import {
 } from 'react-native-device-info';
 import md5 from 'md5';
 import { useState } from 'react';
+import { LOCALES, DYNAMIC_STRINGS } from '../constants';
 
 type FaceType =
   | 'enroll'
@@ -38,6 +43,8 @@ type FaceType =
 export default function Screen() {
   const [isInitialized, setIsInitialized] = useState(false);
   const [isEnabledVocal, setIsEnabledVocal] = useState(false);
+  const [isDynamicStringsEnabled, setIsDynamicStringsEnabled] = useState(false);
+  const [localization, setLocalization] = useState<Locale>('default');
 
   async function onInitialize() {
     const clientInfo = getSystemName();
@@ -111,6 +118,29 @@ export default function Screen() {
     setIsEnabledVocal(enabled);
   }
 
+  const onLocale = () => {
+    const localeIndex = Math.floor(Math.random() * (LOCALES.length - 2));
+    const value = LOCALES.filter((l) => l !== localization)[localeIndex]!;
+
+    setLocalization(value);
+    setLocale(value);
+  };
+
+  const onResetLocale = () => {
+    setLocalization('default');
+    setLocale('default');
+  };
+
+  const onDynamicStrings = () => {
+    if (isDynamicStringsEnabled) {
+      setIsDynamicStringsEnabled(false);
+      resetDynamicStrings();
+    } else {
+      setIsDynamicStringsEnabled(true);
+      setDynamicStrings(DYNAMIC_STRINGS);
+    }
+  };
+
   const opacityStyle = { opacity: isInitialized ? 1 : 0.5 };
   const vocalStyle = {
     backgroundColor: isEnabledVocal ? 'green' : 'red',
@@ -171,8 +201,29 @@ export default function Screen() {
         <Text style={styles.text}>Photo Scan</Text>
       </TouchableOpacity>
 
+      <TouchableOpacity
+        disabled={!isInitialized}
+        style={[styles.button, opacityStyle]}
+        onPress={onLocale}
+        onLongPress={onResetLocale}
+      >
+        <Text style={styles.text}>Localization: {localization}</Text>
+      </TouchableOpacity>
+
       <TouchableOpacity onPress={vocal} style={[styles.button, vocalStyle]}>
         <Text style={styles.text}>Vocal {isEnabledVocal ? 'On' : 'Off'}</Text>
+      </TouchableOpacity>
+
+      <TouchableOpacity
+        disabled={!isInitialized}
+        style={[styles.button, opacityStyle]}
+        onPress={onDynamicStrings}
+      >
+        <Text style={styles.text}>
+          {isDynamicStringsEnabled
+            ? 'Reset Dynamic Strings'
+            : 'Set Dynamic Strings'}
+        </Text>
       </TouchableOpacity>
     </FaceView>
   );