+++ ios/Theme/Button.swift
@@ -1,35 +1,48 @@
 import Foundation
 
-public class Button {
-  private let theme: NSDictionary?
+public class Button: ViewStyle {
+  private static let KEY: String = "guidance"
+  private let target: NSDictionary?
+  private let font: Font
   private let color: Color
 
-  init(theme: NSDictionary?) {
-    self.theme = Style().getTarget(theme, key: "button")
+  init(target: NSDictionary?) {
+    self.target = Style().getTarget(target, key: Button.KEY)
+    self.font = Font()
     self.color = Color()
+    
+    super.init(key: Button.KEY)
+  }
+  
+  override public func getCornerRadius() -> Int32 {
+    return super.getCornerRadius(-1)
+  }
+  
+  public func getFont() -> UIFont {
+    return self.font.getTypography(theme: self.target, key: "font")
   }
 
   public func getBackgroundNormalColor() -> UIColor {
-    return self.color.getColor(self.theme, key: "backgroundNormalColor", defaultColor: "#026ff4")
+    return self.color.getColor(self.target, key: "backgroundNormalColor", defaultColor: "#026ff4")
   }
 
   public func getBackgroundDisabledColor() -> UIColor {
-    return self.color.getColor(self.theme, key: "backgroundDisabledColor", defaultColor: "#b3d4fc")
+    return self.color.getColor(self.target, key: "backgroundDisabledColor", defaultColor: "#b3d4fc")
   }
 
   public func getBackgroundHighlightColor() -> UIColor {
-    return self.color.getColor(self.theme, key: "backgroundHighlightColor", defaultColor: "#0264dc")
+    return self.color.getColor(self.target, key: "backgroundHighlightColor", defaultColor: "#0264dc")
   }
 
   public func getTextNormalColor() -> UIColor {
-    return self.color.getColor(self.theme, key: "textNormalColor")
+    return self.color.getColor(self.target, key: "textNormalColor")
   }
 
   public func getTextDisabledColor() -> UIColor {
-    return self.color.getColor(self.theme, key: "textDisabledColor")
+    return self.color.getColor(self.target, key: "textDisabledColor")
   }
 
   public func getTextHighlightColor() -> UIColor {
-    return self.color.getColor(self.theme, key: "textHighlightColor")
+    return self.color.getColor(self.target, key: "textHighlightColor")
   }
 }