+++ ios/Theme/IdScan.swift
@@ -1,16 +1,27 @@
 public class IdScan {
+  private let target: NSDictionary?
+  private let font: Font
   private let button: Button
   private let selectionScreen: SelectionScreen
   private let captureScreen: CaptureScreen
   private let reviewScreen: ReviewScreen
 
   init() {
-    let theme = Style().getTarget("idScan")
+    self.target = Style().getTarget("idScan")
 
-    self.button = Button(theme: theme)
-    self.selectionScreen = SelectionScreen(theme: theme)
-    self.captureScreen = CaptureScreen(theme: theme)
-    self.reviewScreen = ReviewScreen(theme: theme)
+    self.font = Font()
+    self.button = Button(target: self.target)
+    self.selectionScreen = SelectionScreen(target: self.target)
+    self.captureScreen = CaptureScreen(target: self.target)
+    self.reviewScreen = ReviewScreen(target: self.target)
+  }
+  
+  public func getHeaderFont() -> UIFont {
+    return self.font.getTypography(theme: self.target, key: "headerFont")
+  }
+  
+  public func getSubtextFont() -> UIFont {
+    return self.font.getTypography(theme: self.target, key: "subtextFont")
   }
 
   public func getButton() -> Button {